Update login wording to use preferred 'sign in'
