+++ src/MapsApp.Android/MapsApp.Android.csproj
@@ -73,7 +73,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="LoginPageRenderer.cs" />
+    <Compile Include="SignInPageRenderer.cs" />
     <Compile Include="MainActivity.cs" />
     <Compile Include="Resources\Resource.Designer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
@@ -106,7 +106,7 @@
     <AndroidResource Include="Resources\Drawable\chooseusermap.png" />
   </ItemGroup>
   <ItemGroup>
-    <AndroidResource Include="Resources\Drawable\login_logout.png" />
+    <AndroidResource Include="Resources\Drawable\sign_in_out.png" />
   </ItemGroup>
   <ItemGroup>
     <AndroidResource Include="Resources\Drawable\menu.png" />
