+++ src/MapsApp.Android/Resources/drawable/sign_in_out.png
