+++ src/MapsApp.UWP/sign_in_out.png
