+++ src/MapsApp.WPF/Images/sign_in_out.png
