+++ src/MapsApp.iOS/Resources/Media.xcassets/sign_in_out.imageset/Contents.json
@@ -19,9 +19,7 @@
       "idiom": "universal"
     }
   ],
-  "properties": {
-    "template-rendering-intent": ""
-  },
+  "properties": {},
   "info": {
     "version": 1,
     "author": ""
