+++ src/MapsApp.iOS/Resources/Media.xcassets/sign_in_out.imageset/iOS8_SingleUserDetails30.png
