+++ src/MapsApp.iOS/SignInPageRenderer.cs
@@ -14,38 +14,34 @@
 //  *   limitations under the License.
 //  ******************************************************************************/
 
-using Android.App;
+using Esri.ArcGISRuntime.ExampleApps.MapsApp.Helpers;
+using Esri.ArcGISRuntime.ExampleApps.MapsApp.iOS;
+using Esri.ArcGISRuntime.ExampleApps.MapsApp.Xamarin;
 using Esri.ArcGISRuntime.Security;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using Xamarin.Forms;
-using Xamarin.Forms.Platform.Android;
 using Xamarin.Auth;
-using Esri.ArcGISRuntime.ExampleApps.MapsApp.Helpers;
-using Esri.ArcGISRuntime.ExampleApps.MapsApp.Xamarin;
+using Xamarin.Forms;
+using Xamarin.Forms.Platform.iOS;
 
-[assembly: ExportRenderer(typeof(StartPage), typeof(Esri.ArcGISRuntime.ExampleApps.MapsApp.Android.LoginPageRenderer))]
-namespace Esri.ArcGISRuntime.ExampleApps.MapsApp.Android
+[assembly: ExportRenderer(typeof(StartPage), typeof(SignInPageRenderer))]
+namespace Esri.ArcGISRuntime.ExampleApps.MapsApp.iOS
 {
-    public class LoginPageRenderer : PageRenderer, IOAuthAuthorizeHandler
+    public class SignInPageRenderer : PageRenderer, IOAuthAuthorizeHandler
     {
         // ctor
-        public LoginPageRenderer()
+        public SignInPageRenderer()
         {
             // Set the OAuth authorization handler to this class (Implements IOAuthAuthorize interface)
             AuthenticationManager.Current.OAuthAuthorizeHandler = this;
         }
 
         #region IOAuthAuthorizationHandler implementation
-        // IOAuthAuthorizeHandler.AuthorizeAsync implementation
         public Task<IDictionary<string, string>> AuthorizeAsync(Uri serviceUri, Uri authorizeUri, Uri callbackUri)
         {
             // Create a task completion source
             var taskCompletionSource = new TaskCompletionSource<IDictionary<string, string>>();
-
-            // Get the current Android Activity
-            var activity = this.Context as Activity;
 
             // Create a new Xamarin.Auth.OAuth2Authenticator using the information passed in
             var authenticator = new OAuth2Authenticator(
@@ -57,7 +53,7 @@ namespace Esri.ArcGISRuntime.ExampleApps.MapsApp.Android
             // Allow the user to cancel the OAuth attempt
             authenticator.AllowCancel = true;
 
-            // Hide errors
+            // hide error messages (Xamarin bug)
             authenticator.ShowErrors = false;
 
             // Define a handler for the OAuth2Authenticator.Completed event
@@ -65,6 +61,9 @@ namespace Esri.ArcGISRuntime.ExampleApps.MapsApp.Android
             {
                 try
                 {
+                    // Dismiss the OAuth UI when complete
+                    this.DismissViewController(true, null);
+
                     // Throw an exception if the user could not be authenticated
                     if (!authArgs.IsAuthenticated)
                     {
@@ -80,19 +79,17 @@ namespace Esri.ArcGISRuntime.ExampleApps.MapsApp.Android
                 catch (Exception ex)
                 {
                     // If authentication failed, set the exception on the TaskCompletionSource
-                    taskCompletionSource.TrySetException(ex);
-                }
-                finally
-                {
-                    // Dismiss the OAuth login
-                    activity.FinishActivity(99);
+                    taskCompletionSource.SetException(ex);
                 }
             };
 
             // If an error was encountered when authenticating, set the exception on the TaskCompletionSource
             authenticator.Error += (sndr, errArgs) =>
             {
-                // If the user cancels, the Error event is raised but there is no exception
+                // Dismiss the OAuth UI when complete
+                this.DismissViewController(true, null);
+
+                // If the user cancels, the Error event is raised but there is no exception ... best to check first
                 if (errArgs.Exception != null)
                 {
                     taskCompletionSource.TrySetException(errArgs.Exception);
@@ -101,13 +98,13 @@ namespace Esri.ArcGISRuntime.ExampleApps.MapsApp.Android
                 {
                     authenticator.OnCancelled();
                 }
-
-                activity.FinishActivity(99);
             };
 
-            // Present the OAuth UI so the user can enter user name and password
-            var intent = authenticator.GetUI(activity);
-            activity.StartActivityForResult(intent, 99);
+            // Present the OAuth UI (on the app's UI thread) so the user can enter user name and password
+            InvokeOnMainThread(() =>
+            {
+                this.PresentViewController(authenticator.GetUI(), true, null);
+            });
 
             // Return completion source task so the caller can await completion
             return taskCompletionSource.Task;
